+++ .github/workflows/rust.yml
@@ -69,12 +69,12 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: build
-          args: --release --all-features
+          args: --release
 
       - name: Test
         uses: GabrielBB/xvfb-action@v1
         with:
-          run: cargo test
+          run: cargo test --features=test-ui-replay
 
       - uses: actions/upload-artifact@v2
         with:
