+++ src/lib.rs
@@ -1,0 +1,19 @@
+#[macro_use]
+extern crate bitfield;
+
+mod backend;
+mod capture;
+pub mod decoder;
+mod expander;
+mod file_vec;
+mod hybrid_index;
+mod id;
+pub mod model;
+pub mod row_data;
+mod tree_list_model;
+pub mod ui;
+mod usb;
+mod vec_map;
+
+#[cfg(any(feature="test-ui-replay", feature="record-ui-test"))]
+pub mod record_ui;
