+++ src/record_ui.rs
@@ -42,7 +42,7 @@ pub struct Recording {
     action_log: File,
     #[cfg(feature="record-ui-test")]
     output_log: File,
-    #[cfg(test)]
+    #[cfg(feature="test-ui-replay")]
     output_log: Option<File>,
     view_items: HashMap<String, Vec<String>>,
 }
@@ -66,13 +66,13 @@ impl Recording {
                 .truncate(true)
                 .open("output.txt")
                 .expect("Failed to open UI output log file"),
-            #[cfg(test)]
+            #[cfg(feature="test-ui-replay")]
             output_log: None,
             view_items: HashMap::new(),
         }
     }
 
-    #[cfg(test)]
+    #[cfg(feature="test-ui-replay")]
     pub fn set_output(&mut self, file: File) {
         self.output_log = Some(file)
     }
@@ -93,7 +93,7 @@ impl Recording {
     fn log_output(&mut self, string: String) {
         #[cfg(feature="record-ui-test")]
         let output_log = &mut self.output_log;
-        #[cfg(test)]
+        #[cfg(feature="test-ui-replay")]
         let output_log = self.output_log
             .as_mut()
             .expect("Recording has no output file set");
